Additional improvements for no OFED support.

+++ common_config.sh
@@ -2,51 +2,72 @@
 source ./config.dat
 
 function setdevlink() {
-# setdevlink [mlx_id] [controller IP] [type=full/crypto]
-  CMLXID=$1
+# setdevlink [PCIDEV] [controller IP] [type=full/crypto]
+  set -x
+  PCIDEV="0000:$1"
   CTRLR=$2
   TYPE=$3
+
+  if [[ "$CTRLR" == "$LCTRLR" ]]; then
+      HOST=$LHOST
+      HPCIDEV="0000:$LPCIDEV"
+  else
+      HOST=$RHOST
+      HPCIDEV="0000:$RPCIDEV"
+  fi
+  ssh -x $HOST <<EOF
+  set -x
+  if [ -e /sys/bus/pci/devices/${HPCIDEV}/mlx5_num_vfs ]; then 
+      echo 0 > /sys/bus/pci/devices/${HPCIDEV}/mlx5_num_vfs
+  fi
+EOF
   
   ssh -x $CTRLR /bin/bash <<EOF
      set -x #echo on
 
      ip x s f
      ip x p f
-     CIF=\`ls /sys/class/infiniband/$CMLXID/device/net\`
+     CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
      CIF=\${CIF/\//}
-     PCIDEV=\`mst status -v | grep \$CIF | awk '{print \$3}'\`
-     PCIDEV="0000:\$PCIDEV"
-     echo 0 > /sys/class/infiniband/${CMLXID}/device/mlx5_num_vfs
-     devlink dev eswitch set pci/\${PCIDEV} mode legacy
-     echo dmfs > /sys/bus/pci/devices/\${PCIDEV}/net/\${CIF}/compat/devlink/steering_mode
-     if [[ $TYPE == "full" ]]; then
-       echo full > /sys/class/net/\${CIF}/compat/devlink/ipsec_mode
-       devlink dev eswitch set pci/\${PCIDEV} mode switchdev
+     if [ -e /sys/bus/pci/devices/${PCIDEV}/net/\${CIF}/compat/devlink/steering_mode ]; then
+       devlink dev eswitch set pci/${PCIDEV} mode legacy
+       sleep 1
+       echo dmfs > /sys/bus/pci/devices/${PCIDEV}/net/\${CIF}/compat/devlink/steering_mode
+       if [[ $TYPE == "full" ]]; then
+           echo full > /sys/class/net/\${CIF}/compat/devlink/ipsec_mode
+       fi
+       devlink dev eswitch set pci/${PCIDEV} mode switchdev
      fi
-     service openvswitch stop
+     grep Ubuntu /etc/os-release >& /dev/null
+     if [[ \$? == 0 ]]; then
+       service openvswitch stop
+     else
+       service openvswitch-switch stop
+     fi
 EOF
 }
 
 function set_host_vfs() {
-# setvfs [host IP] [host device mlx id]
+# setvfs [host IP] [host PCI device]
     
   HOST=$1
-  MLXID=$2
+#  MLXID=$2
+  PCIDEV="0000:$2"
   
   ssh -x $HOST /bin/bash <<EOF
   service NetworkManager stop
   set -x
-  echo "$VFS > /sys/class/infiniband/${MLXID}/device/mlx5_num_vfs"
-  echo "$VFS" > /sys/class/infiniband/${MLXID}/device/mlx5_num_vfs
-  
+  echo "$VFS > /sys/bus/pci/devices/${PCIDEV}/mlx5_num_vfs"
+  echo "$VFS" > /sys/bus/pci/devices/${PCIDEV}/mlx5_num_vfs
 EOF
 }
 
 function set_ovs() {
-# set_ovs [controller IP] [controller MLXID] [Offload=on/off]
+# set_ovs [controller IP] [controller PCIDEV] [Offload=on/off]
     
   CTRLRIP=$1
-  CMLXID=$2
+ # CMLXID=$2
+  PCIDEV="0000:$2"
   OFFLOAD=${3:-"on"}
 
   if [[ "$OFFLOAD" == "on" ]]; then
@@ -58,10 +79,10 @@ function set_ovs() {
   ssh $CTRLRIP /bin/bash << EOF
     set -x
     service NetworkManager stop
-    CIF=\`ls /sys/class/infiniband/$CMLXID/device/net\`
+    CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
     CIF=\${CIF/\//}
     PF0=\$CIF
-    VF0=\`ls /sys/class/infiniband/$CMLXID/device/virtfn0/net/ 2> /dev/null\`
+    VF0=\`ls /sys/bus/pci/devices/${PCIDEV}/virtfn0/net/ 2> /dev/null\`
     if [[ "\$VF0" == "" ]]; then
       # Controller is BF
        case "\$PF0" in
@@ -78,23 +99,30 @@ function set_ovs() {
   
     ethtool -K \$PF0 hw-tc-offload $OFFLOAD
     ifconfig \$PF0 $OUTER_LOCAL_IP up
-    service openvswitch start
+    grep Ubuntu /etc/os-release > /dev/null
+    if [[ \$? == 0 ]]; then
+       OVS="openvswitch"
+    else
+       OVS="openvswitch-switch"
+    fi
+    service \$OVS start
     ovs-vsctl del-br $OVSBR
     ovs-vsctl add-br $OVSBR
     ovs-vsctl add-port $OVSBR \$VF0_REP
 	
    ovs-vsctl set Open_vSwitch . other_config:hw-offload=$OFL
-   service openvswitch restart
+   service \$OVS restart
    ifconfig $OVSBR up
    ovs-vsctl show
 EOF
 }
 
 function set_vxlan_ovs() {
-# set_vxlan_ovs [controller IP] [controller MLXID] [Local GW IP] [Remote GW IP] [VFIP0] [Offload=on/off]
+# set_vxlan_ovs [controller IP] [controller PCIDEV] [Local GW IP] [Remote GW IP] [VFIP0] [Offload=on/off]
     
   CTRLRIP=$1
-  CMLXID=$2
+#  CMLXID=$2
+  PCIDEV="0000:$2"
   OUTER_LOCAL_IP=$3
   OUTER_REMOTE_IP=$4
   VF0IP=$5
@@ -109,10 +137,10 @@ function set_vxlan_ovs() {
   ssh $CTRLRIP /bin/bash << EOF
     set -x
     service NetworkManager stop
-    CIF=\`ls /sys/class/infiniband/$CMLXID/device/net\`
+    CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
     CIF=\${CIF/\//}
     PF0=\$CIF
-    VF0=\`ls /sys/class/infiniband/$CMLXID/device/virtfn0/net/ 2> /dev/null\`
+    VF0=\`ls /sys/bus/pci/devices/${PCIDEV}/virtfn0/net/ 2> /dev/null\`
     if [[ "\$VF0" == "" ]]; then
       # Controller is BF
        case "\$PF0" in
@@ -132,8 +160,20 @@ function set_vxlan_ovs() {
     ethtool -K \$VF0_REP hw-tc-offload $OFFLOAD
     ethtool -K \$PF0 hw-tc-offload $OFFLOAD
     ifconfig \$PF0 $OUTER_LOCAL_IP up
-    service openvswitch start
-    ovs-vsctl del-br $OVSBR
+    grep Ubuntu /etc/os-release > /dev/null
+    if [[ \$? == 0 ]]; then
+       OVS="openvswitch"
+    else
+       OVS="openvswitch-switch"
+    fi
+    service \$OVS start
+    sleep 2
+    BRS=(\`ovs-vsctl show | grep Bridge | awk '{print \$2}'\`)
+    if [ \${#BRS[@]} -gt 0 ]; then
+       for br in \${BRS[@]}; do
+          ovs-vsctl del-br \$br
+       done
+    fi
     ovs-vsctl add-br $OVSBR
     ovs-vsctl add-port $OVSBR \$VF0_REP
     ovs-vsctl add-port $OVSBR vxlan${VXLAN_ID} -- set interface vxlan${VXLAN_ID} type=vxlan \
@@ -141,22 +181,22 @@ function set_vxlan_ovs() {
     options:dst_port=$VXLAN_PORT
 	
    ovs-vsctl set Open_vSwitch . other_config:hw-offload=$OFL
-   service openvswitch restart
+   service \$OVS restart
    ifconfig $OVSBR up
    ovs-vsctl show
 EOF
 }
 
 function add_ovs_vxlan_ports() {
-# add_ovs_vxlan_ports [server IP] [Controller MLX DEVID]
+# add_ovs_vxlan_ports [server IP] [Controller PCIDEV]
     CTRLRIP=$1
-    CMLXID=$2
+    PCIDEV="0000:$2"
 
     ssh $CTRLRIP /bin/bash << EOF
       NOVFS=($(seq 0 $(( $VFS - 1 ))))
-      CIF=\`ls /sys/class/infiniband/$CMLXID/device/net\`
+      CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
       CIF=\${CIF/\//}
-      VF0=\`ls /sys/class/infiniband/$CMLXID/device/virtfn0/net/ 2> /dev/null\`
+      VF0=\`ls /sys/bus/pci/devices/${PCIDEV}/virtfn0/net/ 2> /dev/null\`
       if [[ "\$VF0" == "" ]]; then
         case "\$CIF" in
           p0)
@@ -181,24 +221,25 @@ EOF
 }
 
 function set_host_vf_ips() {
-# set_host_vf_ips [HOST] [MLXID] [VF0IP] [offload]
+# set_host_vf_ips [HOST] [PCIDEV] [VF0IP] [offload]
     
   HOST=$1
-  HMLXID=$2
+  PCIDEV="0000:$2"
   VF0IP=$3
   OFL=${4:-"on"}
   ssh -x $HOST /bin/bash <<EOF
-    #set -x
-    CIF=\`ls /sys/class/infiniband/$HMLXID/device/net\`
+    set -x
+    CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
     CIF=\${CIF/\//}
     IPB=(`echo $VF0IP | sed "s/\./ /g"`)
     IP3B=\${IPB[2]}
-    #set -x
-    ifconfig \${CIF}_0 > /dev/null
+    ip l s \${CIF}_0 >& /dev/null
+# With host-aware IPSec, both PF and representor would show.
     if [[ \$? == 1 ]]; then
-      netdevs=(dummy \$CIF \`ls /sys/class/infiniband/$HMLXID/device/virtfn*/net/\`)
+      netdevs=(dummy \$CIF \`ls /sys/bus/pci/devices/${PCIDEV}/virtfn*/net/\`)
     else
-      netdevs=(\`ls /sys/class/infiniband/$HMLXID/device/virtfn*/net/\`)
+# Transparent IPSec
+      netdevs=(\`ls /sys/bus/pci/devices/${PCIDEV}/virtfn*/net/\`)
     fi
     for i in \$( seq 1 2 \${#netdevs[@]} ); do
       ND=\${netdevs[\$i]}
@@ -208,21 +249,22 @@ function set_host_vf_ips() {
       echo \$cmd
       eval "\$cmd"
       echo \$cmd2
-      eval "\$cmd2"
+#     eval "\$cmd2"
     done
 EOF
 }
 
 function set_host_ips() {
-# set_host_ips [HOST] [MLXID] [VF0IP]
+# set_host_ips [HOST] [PCIDEV] [VF0IP]
     
   HOST=$1
-  HMLXID=$2
+  PCIDEV="0000:$2"
   VF0IP=$3
 
   ssh -x $HOST /bin/bash <<EOF
     #set -x
-    CIF=\`ls /sys/class/infiniband/$HMLXID/device/net\`
+#    CIF=\`ls /sys/class/infiniband/$HMLXID/device/net\`
+    CIF=\`ls /sys/bus/pci/devices/${PCIDEV}/net\`
     CIF=\${CIF/\//}
     IPB=(`echo $VF0IP | sed "s/\./ /g"`)
     IP3B=\${IPB[2]}
@@ -238,15 +280,15 @@ EOF
 
 
 function set_host_vf_mtu() {
-# set_vf_mtu [HOST] [MLXID] [mtu]
+# set_vf_mtu [HOST] [PCIDEV] [mtu]
     
   HOST=$1
-  HMLXID=$2
+  PCIDEV="0000:$2"
   HMTU=${3:-$MTU}
   ssh -x $HOST /bin/bash <<EOF
     #set -x
 
-    netdevs=(\`ls /sys/class/infiniband/$HMLXID/device/virtfn*/net/\`)
+    netdevs=(\`ls /sys/bus/pci/devices/${PCIDEV}/virtfn*/net/\`)
     for i in \$( seq 1 2 \${#netdevs[@]} ); do
       ND=\${netdevs[\$i]}
       cmd="ip link set dev \$ND mtu $HMTU"
